Renamed serial module to console. This module will from now take care of all the different ways to write to console
