+++ Cargo.toml
@@ -38,5 +38,4 @@ default = ["panic-over-serial"]
 itm = []
 panic-over-itm = ["itm"]
 abort-on-panic = ["cortex-m-rt/abort-on-panic"]
-serial = []
-panic-over-serial = ["serial"]
+panic-over-serial = []
